// interface/api/prayerTimeOptions.ts
export interface PrayerTimeOptions {
	date?: string;
	latitude?: number;
	longitude?: number;
	city?: string;
	country?: string;
	state?: string;
	year?: number;
	month?: number;
	method?: number;
	shafaq?: string;
	tune?: string;
	school?: number;
	midnightMode?: number;
	timezonestring?: string;
	latitudeAdjustmentMethod?: number;
	adjustment?: number;
	iso8601?: boolean;
	address?: string;
}